create basic bmp info table
